refactor(guardrails): split Alice WonderFence module + tests by concern, drop in-repo doc
Addresses two PR #26901 blockers:

1. **Size-gate CI**: `alice_wonderfence.py` (+627 LOC) and the monolithic test
   file (+1011 LOC) tripped the 500-added-LOC threshold. Both are split along
   separation-of-concerns boundaries — no behavioral changes, only relocation
   and import rewiring. Largest resulting file is 496 LOC.

   Production split:
   - exceptions.py — WonderFenceMissingSecrets, WonderFenceBlockedError
   - client_cache.py — SDK lazy import + LRU client cache helper
   - credentials.py — api_key/app_id resolution + request-scoped stash bridge
   - processing.py — analysis context build, text extract, action dispatch
   - alice_wonderfence.py — WonderFenceGuardrail class (orchestrator)

   Test split (under tests/.../alice_wonderfence/):
   - conftest.py — shared SDK-stub + guardrail-factory fixtures
   - test_credentials.py — resolver precedence + override-flag tests
   - test_client_cache.py — LRU cache + initialization + missing-SDK tests
   - test_apply_guardrail.py — BLOCK/MASK/DETECT/NO_ACTION + fail modes
   - test_post_call_bridge.py — logging_obj stash + sibling fallback

2. **Maintainer request**: drop docs/my-website/docs/proxy/guardrails/
   alice_wonderfence.md from this repo per CLAUDE.md (docs live in
   BerriAI/litellm-docs). The page has been ported to litellm-docs in
   https://github.com/BerriAI/litellm-docs/pull/176.
